+++ R/data_read.R
@@ -1,4 +1,3 @@
-
 #' @title Read (import) data files from various sources
 #' @name data_read
 #'
@@ -109,7 +108,7 @@ data_read <- function(path,
 
 .extract_zip <- function(path) {
   files <- utils::unzip(path, list = TRUE)
-  files_ext <- sapply(files$Name, .file_ext)
+  files_ext <- vapply(files$Name, .file_ext, FUN.VALUE = character(1L))
 
   supported_filetypes <- c("txt", "csv", "xls", "xlsx", "sav", "por", "dta")
   dest <- files$Name[which(files_ext %in% supported_filetypes)]