+++ R/utils_standardize_center.R
@@ -1,4 +1,3 @@
-
 # helper -----------------------------
 
 
@@ -7,7 +6,7 @@
 .process_std_center <- function(x,
                                 weights,
                                 robust,
-                                verbose,
+                                verbose = TRUE,
                                 reference = NULL,
                                 center = NULL,
                                 scale = NULL) {
@@ -38,7 +37,7 @@
   }
 
   # Get center and scale
-  ref <- .get_center_scale(vals, robust, weights, reference, .center = center, .scale = scale)
+  ref <- .get_center_scale(vals, robust, weights, reference, .center = center, .scale = scale, verbose = verbose)
 
   list(
     vals = vals,
@@ -105,8 +104,8 @@
 
   omit <- switch(remove_na,
     none = logical(nrow(x)),
-    selected = rowSums(sapply(x[select], is.na)) > 0,
-    all = rowSums(sapply(x, is.na)) > 0
+    selected = rowSums(vapply(x[select], is.na, FUN.VALUE = logical(nrow(x)))) > 0,
+    all = rowSums(vapply(x, is.na, FUN.VALUE = logical(nrow(x)))) > 0
   )
   x <- x[!omit, , drop = FALSE]
 
@@ -193,7 +192,7 @@
 
 ## retrieve center and scale information ----
 
-.get_center_scale <- function(x, robust = FALSE, weights = NULL, reference = NULL, .center = NULL, .scale = NULL) {
+.get_center_scale <- function(x, robust = FALSE, weights = NULL, reference = NULL, .center = NULL, .scale = NULL, verbose = TRUE) {
   if (is.null(reference)) reference <- x
 
   # for center(), we have no scale. default to 0
@@ -215,10 +214,12 @@
 
   if (scale == 0) {
     scale <- 1
-    insight::format_warning(sprintf(
-      "%s is 0 - variable not standardized (only scaled).",
-      if (robust) "MAD" else "SD"
-    ))
+    if (verbose) {
+      insight::format_warning(sprintf(
+        "%s is 0 - variable not standardized (only scaled).",
+        if (robust) "MAD" else "SD"
+      ))
+    }
   }
 
   list(center = center, scale = scale)
@@ -282,9 +283,9 @@
 
   if (!force) {
     if (!keep_character) {
-      factors <- sapply(x[select], function(i) is.factor(i) | is.character(i))
+      factors <- vapply(x[select], function(i) is.factor(i) | is.character(i), FUN.VALUE = logical(1L))
     } else {
-      factors <- sapply(x[select], function(i) is.factor(i))
+      factors <- vapply(x[select], function(i) is.factor(i), FUN.VALUE = logical(1L))
     }
     select <- select[!factors]
   }