+++ tests/testthat/test-convert_to_na.R
@@ -1,4 +1,3 @@
-
 data(iris)
 
 test_that("convert_to_na-factor", {