+++ tests/testthat/test-data_match.R
@@ -1,4 +1,3 @@
-# styler: off
 data(efc, package = "datawizard")
 
 test_that("data_match works as expected", {
@@ -158,7 +157,9 @@ test_that("programming with data_filter", {
   # One arg ------------
 
   foo <- function(var) {
-    data_filter(mtcars, filter = {var} > 30)
+    data_filter(mtcars, filter = {
+      var
+    } > 30)
   }
   expect_equal(
     foo("mpg"),
@@ -176,7 +177,11 @@ test_that("programming with data_filter", {
   # Two args -----------
 
   foo <- function(var, var2) {
-    data_filter(mtcars, filter = {var} > 30 & {var2} <= 66)
+    data_filter(mtcars, filter = {
+      var
+    } > 30 & {
+      var2
+    } <= 66)
   }
   expect_equal(
     foo("mpg", "hp"),
@@ -200,7 +205,11 @@ test_that("programming with data_filter in global env", {
     data_filter(mtcars, "mpg >= 30 & hp <= 66")
   )
   expect_equal(
-    data_filter(mtcars, {var} > 30 & {var2} <= 66),
+    data_filter(mtcars, {
+      var
+    } > 30 & {
+      var2
+    } <= 66),
     data_filter(mtcars, "mpg >= 30 & hp <= 66")
   )
 })
@@ -235,11 +244,10 @@ test_that("data_filter programming works with groups", {
   class(expected) <- c("grouped_df", "data.frame")
   attributes(expected)$groups <- attributes(test)$groups
 
-  var = "x"
+  var <- "x"
 
   expect_identical(
     data_filter(test, "{var} == min({var})"),
     expected
   )
 })
-# styler: on